2_transcription_worker/requirements.txt
openai-whisper
pika
sqlalchemy
psycopg2-binary
python-dotenv
setuptools
